+++ notebook/tutorial.clj
@@ -89,7 +89,7 @@
    [org.soulspace.qclojure.domain.state :as qs]
    [org.soulspace.qclojure.domain.gate :as qg]
    [org.soulspace.qclojure.domain.circuit :as qc]
-   [org.soulspace.qclojure.adapter.visualization :as viz]
+   [org.soulspace.qclojure.application.visualization :as viz]
    [org.soulspace.qclojure.adapter.visualization.svg :as svg]
    [org.soulspace.qclojure.adapter.visualization.html :as html]))